+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\Auth\OAuthController;
 use App\Http\Controllers\SitemapController;
 use App\Livewire\Pages\Account\DashboardPage as AccountDashboardPage;
 use App\Livewire\Pages\Admin\DashboardPage as AdminDashboardPage;
@@ -30,6 +31,12 @@ Route::get('/sitemap.xml', SitemapController::class)->name('sitemap');
 Route::middleware('guest')->group(function () {
     Route::get('/nalog/prijava', LoginPage::class)->name('login');
     Route::get('/nalog/registracija', RegisterPage::class)->name('register');
+    Route::get('/nalog/{provider}/preusmeri', [OAuthController::class, 'redirect'])
+        ->whereIn('provider', ['google', 'facebook'])
+        ->name('oauth.redirect');
+    Route::get('/nalog/{provider}/povratak', [OAuthController::class, 'callback'])
+        ->whereIn('provider', ['google', 'facebook'])
+        ->name('oauth.callback');
     Route::get('/nalog/zaboravljena-lozinka', ForgotPasswordPage::class)->name('password.request');
     Route::get('/nalog/reset-lozinke/{token}', ResetPasswordPage::class)->name('password.reset');
 });